-image upload -static files serving -pdf generation (inovices)

+++ controllers/admin.js
@@ -1,6 +1,8 @@
 const { validationResult } = require('express-validator');
 
 const Product = require('../models/product');
+const errorHandler = require('../utility/error-handler');
+const fileHelper = require('../utility/file');
 
 exports.getAddProduct = (req, res, next) => {
     res.render('admin/edit-product', {
@@ -15,9 +17,25 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
+
+    if (!image) {
+        return res.status(422).render('admin/edit-product', {
+            pageTitle: 'Add Product',
+            path: '/admin/add-product',
+            editing: false,
+            hasError: true,
+            product: {
+                title,
+                price,
+                description
+            },
+            errorMessage: 'Attached file is not an image.',
+            validationErrors: []
+        });
+    }
     const errors = validationResult(req);
 
     if (!errors.isEmpty()) {
@@ -28,7 +46,6 @@ exports.postAddProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title,
-                imageUrl,
                 price,
                 description
             },
@@ -36,6 +53,9 @@ exports.postAddProduct = (req, res, next) => {
             validationErrors: errors.array()
         });
     }
+
+    const imageUrl = image.path;
+
     const product = new Product({
         title,
         price,
@@ -43,14 +63,13 @@ exports.postAddProduct = (req, res, next) => {
         imageUrl,
         userId: req.user
     });
+
     product
         .save()
         .then(() => {
             res.redirect('/admin/products');
         })
-        .catch(err => {
-            console.log(err);
-        });
+        .catch(err => errorHandler(err, next));
 };
 
 exports.getEditProduct = (req, res, next) => {
@@ -74,14 +93,14 @@ exports.getEditProduct = (req, res, next) => {
                 validationErrors: []
             });
         })
-        .catch(err => console.log(err));
+        .catch(err => errorHandler(err, next));
 };
 
 exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
     const price = req.body.price;
+    const image = req.file;
     const description = req.body.description;
     const errors = validationResult(req);
 
@@ -93,7 +112,6 @@ exports.postEditProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title,
-                imageUrl,
                 price,
                 description,
                 _id: prodId
@@ -111,14 +129,16 @@ exports.postEditProduct = (req, res, next) => {
 
             product.title = title;
             product.price = price;
-            product.imageUrl = imageUrl;
             product.description = description;
+            if (image) {
+                product.imageUrl = image.path;
+            }
             return product.save();
         })
         .then(result => {
             res.redirect('/admin/products');
         })
-        .catch(err => console.log(err));
+        .catch(err => errorHandler(err, next));
 };
 
 exports.getProducts = (req, res, next) => {
@@ -130,15 +150,23 @@ exports.getProducts = (req, res, next) => {
                 path: '/admin/products'
             });
         })
-        .catch(err => console.log(err));
+        .catch(err => errorHandler(err, next));
 };
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
+    Product.findById(prodId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('Product not found.'));
+            }
+
+            fileHelper.deleteFile(product.imageUrl);
+
+            return Product.deleteOne({ _id: prodId, userId: req.user._id })
+        })
         .then(() => {
-            console.log('DESTROYED PRODUCT');
             res.redirect('/admin/products');
         })
-        .catch(err => console.log(err));
+        .catch(err => errorHandler(err, next));
 };
